requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pytest==8.0.0
alembic==1.13.1
httpx==0.27.0
 
# Settings & validation
pydantic-settings
email-validator
 
# Authentication
python-jose[cryptography]
passlib[bcrypt]
python-multipart
 
# Rate limiting
slowapi
